refactor(web): extract DiscordIcon, BrawlhallaLinkRow, SignedInState from account page
